+++ sema4ai/vscode-client/src/protocols.ts
@@ -70,6 +70,11 @@ export interface RobotTemplate {
     description: string;
 }
 
+export interface ActionTemplate {
+    name: string;
+    description: string;
+}
+
 export interface WorkItem {
     name: string;
     json_path: string;