+++ docs/changelog.md
@@ -5,7 +5,7 @@ Unreleased
     - Logs will show local variables (ROBOT_ROOT is no longer set).
     - ROBOT_ARTIFACTS set to a better place for the log.html.
     - If using `sema4ai-actions 0.7.0` or newer, the results of running an action are printed to the terminal.
-
+- `Sema4.ai: Create Action Package` allows to specify a template to use now (for Action Server v0.10.0 or newer)
 
 New in 1.22.4 (2024-05-22)
 -----------------------------